Changes to the sensors, client and server modules.
Sensors are now simulated separately, which allows simulating only entries or only exits or both at a time. This way they are more true to a real life scenario, since each sensor would run on different hardware.
Removed some unused functions in the sensors, client and server modules. These now only generate data, send it/receive it and update it.
Data should be stored and processed in the raspbpi module.